+++ tomcat/src/main/java/com/techcourse/controller/LoginRequestController.java
@@ -10,12 +10,11 @@ import org.apache.catalina.Session;
 import org.apache.catalina.SessionManager;
 import org.apache.coyote.HttpRequest;
 import org.apache.coyote.HttpResponse;
-import org.apache.coyote.http11.AbstractRequestHandler;
 import org.apache.coyote.http11.MimeType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class LoginRequestController extends AbstractRequestHandler {
+public class LoginRequestController extends AbstractRequestController {
 
     private static final Logger log = LoggerFactory.getLogger(LoginRequestController.class);
     private static final String SUCCESS_LOGIN_REDIRECTION_PATH = "/index.html";
