+++ tomcat/src/main/java/com/techcourse/controller/SignupRequestController.java
@@ -10,10 +10,9 @@ import org.apache.catalina.Session;
 import org.apache.catalina.SessionManager;
 import org.apache.coyote.HttpRequest;
 import org.apache.coyote.HttpResponse;
-import org.apache.coyote.http11.AbstractRequestHandler;
 import org.apache.coyote.http11.MimeType;
 
-public class SignupRequestController extends AbstractRequestHandler {
+public class SignupRequestController extends AbstractRequestController {
 
     private static final String REDIRECTION_PATH = "/index.html";
     private static final String ACCOUNT_KEY = "account";
